+++ app-backend/batch/src/main/scala/groundwork/CreateTaskGrid.scala
@@ -69,7 +69,6 @@ class CreateTaskGrid(
           .update(
             annotationProject.copy(
               aoi = footprint,
-              ready = true,
               taskSizeMeters = Some(taskSizeMeters)
             ),
             annotationProject.id