+++ app-backend/db/src/main/scala/AnnotationProjectDao.scala
@@ -21,7 +21,7 @@ object AnnotationProjectDao
     SELECT
       id, created_at, owner, name, project_type, task_size_meters,
       task_size_pixels, aoi, labelers_team_id, validators_team_id,
-      project_id, ready
+      project_id, status
     FROM
   """ ++ tableF
 
@@ -129,13 +129,13 @@ object AnnotationProjectDao
   ): ConnectionIO[AnnotationProject.WithRelated] = {
     val projectInsert = (fr"INSERT INTO" ++ tableF ++ fr"""
       (id, created_at, owner, name, project_type, task_size_pixels,
-       aoi, labelers_team_id, validators_team_id, project_id, ready)
+       aoi, labelers_team_id, validators_team_id, project_id, status)
     VALUES
       (uuid_generate_v4(), now(), ${user.id}, ${newAnnotationProject.name},
        ${newAnnotationProject.projectType}, ${newAnnotationProject.taskSizePixels},
        ${newAnnotationProject.aoi}, ${newAnnotationProject.labelersTeamId},
        ${newAnnotationProject.validatorsTeamId},
-       ${newAnnotationProject.projectId}, ${newAnnotationProject.ready})
+       ${newAnnotationProject.projectId}, ${newAnnotationProject.status})
     """).update.withUniqueGeneratedKeys[AnnotationProject](
       "id",
       "created_at",
@@ -148,7 +148,7 @@ object AnnotationProjectDao
       "labelers_team_id",
       "validators_team_id",
       "project_id",
-      "ready"
+      "status"
     )
 
     for {
@@ -230,7 +230,7 @@ object AnnotationProjectDao
       validators_team_id = ${project.validatorsTeamId},
       task_size_meters= ${project.taskSizeMeters},
       aoi = ${project.aoi},
-      ready = ${project.ready}
+      status = ${project.status}
     WHERE
       id = $id
     """).update.run;